+++ R/make-file.R
@@ -89,7 +89,7 @@ make_file <- function(group = NULL,
     }
 
     if (length(deleteVar) != 0) {
-      msgWarn <- "Some columns aren't defined in FILGRUPPE. They are now deleted"
+      msgWarn <- "OBS! Some columns aren't defined in FILGRUPPE. They are now deleted"
       is_verbose(x = msgWarn, type = "warning")
       deleteVar <- paste(deleteVar, collapse = ", ")
       is_verbose(deleteVar, "Deleted column(s):", type = "message")
@@ -154,7 +154,7 @@ read_raw <- function(group = NULL,
                      save = FALSE,
                      year = NULL) {
 
-  lifecycle::deprecate_warn("0.0.9", "read_raw()", "make_file()")
+  lifecycle::deprecate_stop("0.0.9", "read_raw()", "make_file()")
   ## make_file(group, koblid, aggregate, save, year)
 }
 
@@ -162,12 +162,12 @@ read_raw <- function(group = NULL,
 #' @rdname read_raw
 lesraw <- function(group = NULL,
                    koblid = NULL,
-                     aggregate = getOption("orgdata.aggregate"),
-                     save = FALSE,
-                     year = NULL) {
+                   aggregate = getOption("orgdata.aggregate"),
+                   save = FALSE,
+                   year = NULL) {
 
-  lifecycle::deprecate_warn("0.0.9", "lesraw()", "lag_fil()")
-  make_file(group, koblid, aggregate, save, year)
+  lifecycle::deprecate_stop("0.0.9", "lesraw()", "lag_fil()")
+  ## make_file(group, koblid, aggregate, save, year)
 }