+++ R/read-file.R
@@ -1,8 +1,16 @@
 #' Read Data File
 #' @description Read rawdata either using `FILID` value or complete file path. It uses the [find_data()] generic method.
 #'    For a \code{ .csv } file, [data.table::fread()] is used and all other arguments
-#'    for \code{fread} function can be used. For a \code{ xlsx } or \code{ .xls } file
+#'    for \code{fread} function can be used. For a \code{ .xlsx } or \code{ .xls } file
 #'    [readxl::read_excel()] function and all of its arguments.
+#' @description Nevertheless, some most used arguments are standardized for `read_file()` and there are:
+#' \itemize{
+#'   \item `nrows` to display maksimum numbers to read
+#'   \item `header` FALSE to give default columnames as `V1`, `V2` etc
+#'   \item `skip` a specific number of raws before reading the data
+#'   \item `trimws` to trim leading and trailing whitespace
+#'   \item `na` for character value to be interpreted as `NA`
+#' }
 #' @param file Use FILID or a complete path of a filename
 #' @param ... All other arguments to be passed related to the file format
 #' @examples
@@ -27,7 +35,13 @@ read_file <- function(file = NULL, ...) {
   ext <- tools::file_ext(file)
   class(file) <- append(class(file), ext)
 
-  find_data(file, ...)
+  dt <- find_data(file, ...)
+
+  if (sum(is.element(class(dt), "data.table")) == 0){
+    data.table::setDT(dt)
+  }
+
+  return(dt)
 }
 
 #' @export